+++ test/BookSingleChain/Dispute.t.sol
@@ -6,7 +6,7 @@ import "./Fixtures.sol";
 contract DisputeTest is TradeFixture {
     using stdStorage for StdStorage;
 
-    uint128 internal tradeIndex;
+    uint256 internal tradeIndex;
     bytes32 internal tradeId;
     address internal relayer = bob;
     address internal disputer = charlie;
@@ -15,7 +15,7 @@ contract DisputeTest is TradeFixture {
     function setUp() public override {
         super.setUp();
         deal(testTokenIn, alice, testAmount);
-        (uint128 _tradeIndex, bytes32 _tradeId) = _requestTrade(
+        (uint256 _tradeIndex, bytes32 _tradeId) = _requestTrade(
             testTokenIn,
             testTokenOut,
             testAmount,
